fixed: separated user controllers & made user.controller.js
